clean config, fix selector initialization
